google_tts/requirements.txt
gTTS
google-cloud-texttospeech